+++ src/main/java/com/frandm/justenoughbackups/client/screen/config/ConfigFieldId.java
@@ -18,6 +18,8 @@ public enum ConfigFieldId {
     PERMISSION_LEVEL("screen.justenoughbackups.config.permission_level", ConfigTab.BACKUPS, ConfigControlType.INT),
     MESSAGE_CHANNEL("screen.justenoughbackups.config.message_channel", ConfigTab.BACKUPS, ConfigControlType.ENUM),
 
+    EXCLUDE_SAFETY_BACKUPS_FROM_INCREMENTAL("screen.justenoughbackups.config.exclude_safety_backups_from_incremental", ConfigTab.BACKUPS, ConfigControlType.BOOLEAN),
+
     AUTOMATIC_BACKUPS_HEADER("screen.justenoughbackups.config.automatic_backups_header", ConfigTab.BACKUPS, ConfigControlType.SECTION_HEADER),
     AUTOMATIC_FULL_ENABLED("screen.justenoughbackups.config.automatic_full_enabled", ConfigTab.BACKUPS, ConfigControlType.BOOLEAN),
     AUTOMATIC_DIFFERENTIAL_ENABLED("screen.justenoughbackups.config.automatic_differential_enabled", ConfigTab.BACKUPS, ConfigControlType.BOOLEAN),